packages/remark-lsx/src/server/routes/list-pages/add-num-condition.spec.ts
import createError from 'http-errors';
import { mock } from 'vitest-mock-extended';

import { addNumCondition } from './add-num-condition';
import type { PageQuery } from './generate-base-query';

describe('addNumCondition() throws 400 http-errors instance ', () => {

  it.concurrent.each`
    offset  | limit
    ${-1}   | ${9}
    ${1}    | ${-9}
  `('when the specified condition is { offset: $offset, limit: $limit }', ({ offset, limit }) => {

    // setup
    const queryMock = mock<PageQuery>();

    // when
    const caller = () => addNumCondition(queryMock, offset, limit);

    // then
    expect(caller).toThrowError(createError(400, "Both specified 'offset' or 'limit must be larger or equal than 0"));
    expect(queryMock.skip).not.toHaveBeenCalledWith();
    expect(queryMock.limit).not.toHaveBeenCalledWith();
  });
});


describe('addNumCondition() set skip and limit with', () => {

  it.concurrent.each`
    offset  | limit                   | expectedSkip   | expectedLimit
    ${1}    | ${Number.MAX_VALUE}     | ${1}           | ${null}
    ${0}    | ${10}                   | ${null}        | ${10}
    ${0}    | ${Number.MAX_VALUE}     | ${null}        | ${null}
    ${NaN}  | ${NaN}                  | ${null}        | ${null}
  `("{ offset: $offset, limit: $limit }'", ({
    offset, limit, expectedSkip, expectedLimit,
  }) => {
    // setup
    const queryMock = mock<PageQuery>();

    // result for q.skip()
    const querySkipResultMock = mock<PageQuery>();
    queryMock.skip.calledWith(expectedSkip).mockImplementation(() => querySkipResultMock);
    // result for q.limit()
    const queryLimitResultMock = mock<PageQuery>();
    queryMock.limit.calledWith(expectedLimit).mockImplementation(() => queryLimitResultMock);
    // result for q.skil().limit()
    const querySkipAndLimitResultMock = mock<PageQuery>();
    querySkipResultMock.limit.calledWith(expectedLimit).mockImplementation(() => querySkipAndLimitResultMock);

    // when
    const result = addNumCondition(queryMock, offset, limit);

    // then
    if (expectedSkip != null) {
      expect(queryMock.skip).toHaveBeenCalledWith(expectedSkip);
      if (expectedLimit != null) {
        expect(querySkipResultMock.limit).toHaveBeenCalledWith(expectedLimit);
        expect(result).toEqual(querySkipAndLimitResultMock); // q.skip().limit()
      }
      else {
        expect(querySkipResultMock.limit).not.toHaveBeenCalled();
        expect(result).toEqual(querySkipResultMock); // q.skil()
      }
    }
    else {
      expect(queryMock.skip).not.toHaveBeenCalled();
      if (expectedLimit != null) {
        expect(queryMock.limit).toHaveBeenCalledWith(expectedLimit);
        expect(result).toEqual(queryLimitResultMock); // q.limit()
      }
      else {
        expect(queryMock.limit).not.toHaveBeenCalled();
        expect(result).toEqual(queryMock); // as-is
      }
    }
  });

});
